move ConditionConfig to config package
