+++ src/main/java/com/teragrep/pth_06/planner/walker/conditions/IndexStatementCondition.java
@@ -46,6 +46,7 @@
 
 package com.teragrep.pth_06.planner.walker.conditions;
 
+import com.teragrep.pth_06.config.ConditionConfig;
 import com.teragrep.pth_06.planner.BloomFilterTempTable;
 import com.teragrep.pth_06.planner.PatternMatch;
 import org.jooq.Condition;
